+++ include/dlaf/eigensolver/gen_to_std/mc.h
@@ -13,6 +13,7 @@
 #include <hpx/include/resource_partitioner.hpp>
 #include <hpx/include/threads.hpp>
 
+#include "dlaf/blas/tile.h"
 #include "dlaf/common/index2d.h"
 #include "dlaf/common/pipeline.h"
 #include "dlaf/common/range2d.h"
@@ -24,7 +25,6 @@
 #include "dlaf/lapack_tile.h"
 #include "dlaf/matrix/distribution.h"
 #include "dlaf/matrix/matrix.h"
-#include "dlaf/tile.h"
 #include "dlaf/util_matrix.h"
 
 namespace dlaf {
