+++ include/dlaf/solver/triangular/impl.h
@@ -14,6 +14,7 @@
 #include <hpx/include/util.hpp>
 #include <hpx/local/future.hpp>
 
+#include "dlaf/blas/tile.h"
 #include "dlaf/common/index2d.h"
 #include "dlaf/common/pipeline.h"
 #include "dlaf/common/vector.h"
@@ -24,7 +25,6 @@
 #include "dlaf/matrix/distribution.h"
 #include "dlaf/matrix/matrix.h"
 #include "dlaf/solver/triangular/api.h"
-#include "dlaf/tile.h"
 #include "dlaf/util_matrix.h"
 
 namespace dlaf {
