+++ test/unit/test_cublas_tile/test_herk.h
@@ -13,7 +13,7 @@
 #include <sstream>
 #include "gtest/gtest.h"
 #include "dlaf/blas/enum_output.h"
-#include "dlaf/cublas_tile.h"
+#include "dlaf/blas/tile.h"
 #include "dlaf/matrix/copy_tile.h"
 #include "dlaf/matrix/tile.h"
 #include "dlaf/memory/memory_view.h"
